optimize tinymcs, event image display

+++ config/event.php
@@ -9,9 +9,9 @@
 return [
     'useCache'   => env('USE_CACHE', false),
     'defaultEventTime'  => env('DEFAULT_EVENT_TIME', '19:00'),
-    'maxImageHeight'    => env('MAX_IMAGE_HEIGHT', 300),
-    'maxImageWidth'     => env('MAX_IMAGE_WIDTH', 600),
-	'maxImageFileSize'  => env('MAX_IMAGE_FILESIZE', 60000), // B
+    'maxImageHeight'    => (int) env('MAX_IMAGE_HEIGHT', 300),
+    'maxImageWidth'     => (int) env('MAX_IMAGE_WIDTH', 600),
+	'maxImageFileSize'  => (int) env('MAX_IMAGE_FILESIZE', 60000),
 	'paginationLimit' 	=> 10,
 	'eventsPaginationLimit'	=> 10,
     'periodicPosition'  => [
